+++ web/rainmaker/dev-packages/egov-uc-dev/src/ui-config/screens/specs/utils/index.js
@@ -456,44 +456,3 @@ export const getTextToLocalMapping = label => {
   }
 };
 
-
-export const downloadChallan = async (queryStr, mode = 'download') => {
- 
- 
-   //let tenantId = get(Challan, "tenantId");
-  // //Added sorting for Challan details
-  // const challanAmtSorted=  orderBy(
-  //   Challan.amount,
-  //   ["amount"],
-  //   ["desc"]);       
-  //   Challan.amount = challanAmtSorted;
-    
-
-  // const queryStr = [
-  //   { key: "key", value:"mcollect-challan" },
-  //   { key: "tenantId", value: tenantId ? tenantId.split(".")[0] : commonConfig.tenantId }
-  // ];
-  console.info("query string=",queryStr);
-  const DOWNLOADRECEIPT = {
-    GET: {
-      URL: "/egov-pdf/download/UC/mcollect-challan",
-      ACTION: "_get",
-    },
-  };
-    try {
-      httpRequest("post", DOWNLOADRECEIPT.GET.URL, DOWNLOADRECEIPT.GET.ACTION, queryStr, { 'Accept': 'application/json' }, { responseType: 'arraybuffer' })
-        .then(res => {
-          res.filestoreIds[0]
-          if (res && res.filestoreIds && res.filestoreIds.length > 0) {
-            res.filestoreIds.map(fileStoreId => {
-              downloadReceiptFromFilestoreID(fileStoreId, mode)
-            })
-          } else {
-            console.log("Error In Acknowledgement form Download");
-          }
-        });
-    } catch (exception) {
-      alert('Some Error Occured while downloading Acknowledgement form!');
-    }
-  
-}
